Fix a couple of type declrations to quite warnings on compaq. JM
git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@1318 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ external/io_grib1/MEL_grib1/gribfuncs.h
@@ -58,7 +58,7 @@ void make_default_grbfn (DATA_INPUT,USER_INPUT,char *);
 int map_lvl (char *,DATA_INPUT *,float *,float *,char *);
 int map_parm (char *,DATA_INPUT *,float *,float *,char *);
 int pack_spatial (long *,unsigned short *,float *,float *,
-		unsigned long **,short,long *,char *);
+		unsigned int **,short,long *,char *);
 void prt_inp_struct (PDS_INPUT *,grid_desc_sec *, BMS_INPUT *,
 		struct BDS_HEAD_INPUT *,float **);
 void upd_child_errmsg (char *,char *);
